+++ scripts/world/WorldSettings.cs
@@ -8,43 +8,42 @@ public class WorldSettings
     public Rect2I Bounds = new Rect2I(-20000, -20000, 40000, 40000);
     public Vector2 WorldCenter { get => Bounds.Position + Bounds.Size / 2; }
     public int PoisosonDiskSamplingIterations = 8;
-    public float NormalizedMinimumCellDistance { get; set; } = 0.8f;
-    public float MinimumCellDistance
+    public double NormalizedMinimumCellDistance { get; set; } = 0.8;
+    public double MinimumCellDistance
     {
-        get => NormalizedMinimumCellDistance * Bounds.Size.Y / 200.0f;
+        get => NormalizedMinimumCellDistance * Bounds.Size.Y / 200.0;
         set => NormalizedMinimumCellDistance = value * 200.0f / Bounds.Size.Y;
     }
 
-    public float NormalizedNoiseFrequency { get; set; } = 1.0f;
-    public float NoiseFrequency
+    public double NormalizedNoiseFrequency { get; set; } = 1.0;
+    public double NoiseFrequency
     {
-        get => NormalizedNoiseFrequency / 4000.0f;
-        set => NormalizedNoiseFrequency = value * 4000.0f;
+        get => NormalizedNoiseFrequency / 4000.0;
+        set => NormalizedNoiseFrequency = value * 4000.0;
     }
     #endregion
 
     #region Tectonic Settings
-    public float ContinentRatio = 0.6f;
-    public float PlateMergeRatio = 0.13f;
-    public float MaxTectonicMovement = 10.0f;
-    public float MaxUplift = 1300.0f;
-    public float UpliftPropagationDecrement = 0.8f;
-    public float UpliftPropagationSharpness = 0.0f;
+    public double ContinentRatio = 0.6;
+    public double PlateMergeRatio = 0.13;
+    public double MaxTectonicMovement = 10.0;
+    public double MaxUplift = 1300.0;
+    public double UpliftPropagationDecrement = 0.8;
+    public double UpliftPropagationSharpness = 0.0;
     #endregion
 
     #region Fluvial Erosion Settings
-    public float ErosionRate = 5f;
-    public float ErosionTimeStep = 0.1f;
-    public float ErosionConvergenceThreshold = 2.0f;
+    public double ErosionRate = 5;
+    public double ErosionTimeStep = 0.1;
+    public double ErosionConvergenceThreshold = 2.0;
     public int MaxErosionIterations = 150;
     #endregion
 
     #region Orbital Settings
-    public float OrbitalInclinationAngle = 20.0f;
-    public float OrbitalRevolutionDays = 64.0f;
-    public float MinutesPerDay = 30.0f;
+    public double OrbitalInclinationAngle = 20.0;
+    public double OrbitalRevolutionDays = 64.0;
+    public double MinutesPerDay = 30.0;
     #endregion
-
 
     public WorldSettings(ulong seed = 234)
     {
